Detach value from inheritance chain if it is explicitly set
Previously for example values in ambience profile "locked" back to
inheriting from general if either ambience/general was changed so
that the values were the same in both profiles.

Now setting <profile>:<value>="something" means it will stay that
way regardless of what happens to defaults / general profile.

It is still possible to reset a value back to inheriting from general
by assigning it to empty string.