[libprofile] Add "libprofile:" prefix to diagnostic logging. Contributes to JB#24365
Warnings from libprofile are most likely to occur during startup
and shutdown and are thus unlikely to be seen by developers of
the applications that use libprofile. When they are spotted in
journal, they get attributed to the process using libprofile and
it can be hard to spot that the problem is related to libprofile.

Add "libprofile:" prefix for all diagnostic logging that happens
from libprofile.

In some special cases the prefix was already explicitly used, remove
those so that the prefix does not show up twice.